kill update and insert field methods

+++ core/modules/file/lib/Drupal/file/Plugin/Field/FieldType/FileFieldItemList.php
@@ -22,55 +22,50 @@ class FileFieldItemList extends ConfigFieldItemList {
   /**
    * {@inheritdoc}
    */
-  public function insert() {
+  public function preSave() {
     parent::insert();
     $entity = $this->getEntity();
 
     // Add a new usage for newly uploaded files.
-    foreach ($this->targetEntities() as $file) {
-      \Drupal::service('file.usage')->add($file, 'file', $entity->entityType(), $entity->id());
-    }
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function update() {
-    parent::update();
-    $entity = $this->getEntity();
-
-    // Get current target file entities and file IDs.
-    $files = $this->targetEntities();
-    $fids = array_keys($files);
-
-    // On new revisions, all files are considered to be a new usage and no
-    // deletion of previous file usages are necessary.
-    if (!empty($entity->original) && $entity->getRevisionId() != $entity->original->getRevisionId()) {
-      foreach ($files as $file) {
+    if ($entity->isNew()) {
+      foreach ($this->targetEntities() as $file) {
         \Drupal::service('file.usage')->add($file, 'file', $entity->entityType(), $entity->id());
       }
-      return;
     }
+    else {
+      // Get current target file entities and file IDs.
+      $files = $this->targetEntities();
+      $fids = array_keys($files);
 
-    // Get the file IDs attached to the field before this update.
-    $field_name = $this->getFieldDefinition()->getName();
-    $original_fids = array();
-    $original_items = $entity->original->getTranslation($this->getLangcode())->$field_name;
-    foreach ($original_items as $item) {
-      $original_fids[] = $item->target_id;
-    }
+      // On new revisions, all files are considered to be a new usage and no
+      // deletion of previous file usages are necessary.
+      if (!empty($entity->original) && $entity->getRevisionId() != $entity->original->getRevisionId()) {
+        foreach ($files as $file) {
+          \Drupal::service('file.usage')->add($file, 'file', $entity->entityType(), $entity->id());
+        }
+        return;
+      }
 
-    // Decrement file usage by 1 for files that were removed from the field.
-    $removed_fids = array_filter(array_diff($original_fids, $fids));
-    $removed_files = \Drupal::entityManager()->getStorageController('file')->loadMultiple($removed_fids);
-    foreach ($removed_files as $file) {
-      \Drupal::service('file.usage')->delete($file, 'file', $entity->entityType(), $entity->id());
-    }
+      // Get the file IDs attached to the field before this update.
+      $field_name = $this->getFieldDefinition()->getName();
+      $original_fids = array();
+      $original_items = $entity->original->getTranslation($this->getLangcode())->$field_name;
+      foreach ($original_items as $item) {
+        $original_fids[] = $item->target_id;
+      }
 
-    // Add new usage entries for newly added files.
-    foreach ($files as $fid => $file) {
-      if (!in_array($fid, $original_fids)) {
-        \Drupal::service('file.usage')->add($file, 'file', $entity->entityType(), $entity->id());
+      // Decrement file usage by 1 for files that were removed from the field.
+      $removed_fids = array_filter(array_diff($original_fids, $fids));
+      $removed_files = \Drupal::entityManager()->getStorageController('file')->loadMultiple($removed_fids);
+      foreach ($removed_files as $file) {
+        \Drupal::service('file.usage')->delete($file, 'file', $entity->entityType(), $entity->id());
+      }
+
+      // Add new usage entries for newly added files.
+      foreach ($files as $fid => $file) {
+        if (!in_array($fid, $original_fids)) {
+          \Drupal::service('file.usage')->add($file, 'file', $entity->entityType(), $entity->id());
+        }
       }
     }
   }
